requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Environment and Configuration
python-dotenv>=1.0.0
pydantic[email]>=2.5.3
pydantic-settings>=2.1.0

# Image Processing (for future ML integration)
pillow>=10.2.0
numpy>=1.26.3

# PDF Report Generation
reportlab>=4.0.8

# Utilities
aiofiles>=23.2.1
